move role_playing_with_function to utils
