--- camel/societies/role_playing_with_functions.py
@@ -1,118 +1,0 @@
-# =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
-# Licensed under the Apache License, Version 2.0 (the “License”);
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an “AS IS” BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
-from typing import List
-
-from colorama import Fore
-
-from camel.agents.chat_agent import FunctionCallingRecord
-from camel.configs import ChatGPTConfig, FunctionCallingConfig
-from camel.functions import MATH_FUNCS, SEARCH_FUNCS, WEATHER_FUNCS
-from camel.societies import RolePlaying
-from camel.types import ModelType
-from camel.utils import print_text_animated
-
-TASK_PROMPT = ("Assuming the current year is 2023, estimate KAUST's "
-               "current age and then add 10 more years to this age, "
-               "and get the current weather of the city where KAUST "
-               "is located.")
-FUNCTION_LIST = [*MATH_FUNCS, *SEARCH_FUNCS, *WEATHER_FUNCS]
-ASSISTANT_ROLE_NAME = "Searcher"
-USER_ROLE_NAME = "Professor"
-
-
-def role_playing_with_function(
-    task_prompt: str = TASK_PROMPT,
-    function_list: list = FUNCTION_LIST,
-    model_type=ModelType.GPT_4_TURBO,
-    chat_turn_limit=10,
-    assistant_role_name: str = ASSISTANT_ROLE_NAME,
-    user_role_name: str = USER_ROLE_NAME,
-) -> None:
-    task_prompt = task_prompt
-    user_model_config = ChatGPTConfig(temperature=0.0)
-
-    function_list = function_list
-    assistant_model_config = FunctionCallingConfig.from_openai_function_list(
-        function_list=function_list,
-        kwargs=dict(temperature=0.0),
-    )
-
-    role_play_session = RolePlaying(
-        assistant_role_name=assistant_role_name,
-        user_role_name=user_role_name,
-        assistant_agent_kwargs=dict(
-            model_type=model_type,
-            model_config=assistant_model_config,
-            function_list=function_list,
-        ),
-        user_agent_kwargs=dict(
-            model_type=model_type,
-            model_config=user_model_config,
-        ),
-        task_prompt=task_prompt,
-        with_task_specify=False,
-    )
-
-    print(
-        Fore.GREEN +
-        f"AI Assistant sys message:\n{role_play_session.assistant_sys_msg}\n")
-    print(Fore.BLUE +
-          f"AI User sys message:\n{role_play_session.user_sys_msg}\n")
-
-    print(Fore.YELLOW + f"Original task prompt:\n{task_prompt}\n")
-    print(
-        Fore.CYAN +
-        f"Specified task prompt:\n{role_play_session.specified_task_prompt}\n")
-    print(Fore.RED + f"Final task prompt:\n{role_play_session.task_prompt}\n")
-
-    n = 0
-    input_assistant_msg, _ = role_play_session.init_chat()
-    while n < chat_turn_limit:
-        n += 1
-        assistant_response, user_response = role_play_session.step(
-            input_assistant_msg)
-
-        if assistant_response.terminated:
-            print(Fore.GREEN +
-                  ("AI Assistant terminated. Reason: "
-                   f"{assistant_response.info['termination_reasons']}."))
-            break
-        if user_response.terminated:
-            print(Fore.GREEN +
-                  ("AI User terminated. "
-                   f"Reason: {user_response.info['termination_reasons']}."))
-            break
-
-        # Print output from the user
-        print_text_animated(Fore.BLUE +
-                            f"AI User:\n\n{user_response.msg.content}\n")
-
-        # Print output from the assistant, including any function
-        # execution information
-        print_text_animated(Fore.GREEN + "AI Assistant:")
-        called_functions: List[
-            FunctionCallingRecord] = assistant_response.info[
-                'called_functions']
-        for func_record in called_functions:
-            print_text_animated(f"{func_record}")
-        print_text_animated(f"{assistant_response.msg.content}\n")
-
-        if "CAMEL_TASK_DONE" in user_response.msg.content:
-            break
-
-        input_assistant_msg = assistant_response.msg
-
-
-if __name__ == "__main__":
-    role_playing_with_function()
