+++ camel/utils/commons.py
@@ -38,6 +38,17 @@ from camel.types import TaskType
 F = TypeVar('F', bound=Callable[..., Any])
 
 
+# Set lazy import
+def get_lazy_imported_functions_module():
+    from camel.functions import MATH_FUNCS, SEARCH_FUNCS, WEATHER_FUNCS
+    return [*MATH_FUNCS, *SEARCH_FUNCS, *WEATHER_FUNCS]
+
+
+def get_lazy_imported_types_module():
+    from camel.types import ModelType
+    return ModelType.GPT_4_TURBO
+
+
 def openai_api_key_required(func: F) -> F:
     r"""Decorator that checks if the OpenAI API key is available in the
     environment variables.
@@ -241,3 +252,129 @@ def check_server_running(server_url: str) -> bool:
 
     # if the port is open, the result should be 0.
     return result == 0
+
+
+def role_playing_with_function(
+    task_prompt: str = ("Assuming the current year is 2023, estimate KAUST's "
+                        "current age and then add 10 more years to this age, "
+                        "and get the current weather of the city where KAUST "
+                        "is located."),
+    function_list: Optional[List] = None,
+    model_type=None,
+    chat_turn_limit=10,
+    assistant_role_name: str = "Searcher",
+    user_role_name: str = "Professor",
+) -> None:
+    r"""Initializes and conducts a `RolePlaying` with `FunctionCallingConfig`
+    session. The function creates an interactive and dynamic role-play session
+    where the AI Assistant and User engage based on the given task, roles, and
+    available functions. It demonstrates the versatility of AI in handling
+    diverse tasks and user interactions within a structured `RolePlaying`
+    framework.
+
+    Args:
+        task_prompt (str): The initial task or scenario description to start
+            the `RolePlaying` session. Defaults to a prompt involving the
+            estimation of KAUST's age and weather information.
+        function_list (list): A list of functions that the agent can utilize
+            during the session. Defaults to a combination of math, search, and
+            weather functions.
+        model_type (ModelType): The type of chatbot model used for both the
+            assistant and the user. Defaults to `GPT-4 Turbo`.
+        chat_turn_limit (int): The maximum number of turns (exchanges) in the
+            chat session. Defaults to 10.
+        assistant_role_name (str): The role name assigned to the AI Assistant.
+            Defaults to 'Searcher'.
+        user_role_name (str): The role name assigned to the User. Defaults to
+            'Professor'.
+
+    Returns:
+        None: This function does not return any value but prints out the
+            session's dialogues and outputs.
+    """
+
+    # Run lazy import
+    if function_list is None:
+        function_list = get_lazy_imported_functions_module()
+    if model_type is None:
+        model_type = get_lazy_imported_types_module()
+
+    from colorama import Fore
+
+    from camel.agents.chat_agent import FunctionCallingRecord
+    from camel.configs import ChatGPTConfig, FunctionCallingConfig
+    from camel.societies import RolePlaying
+
+    task_prompt = task_prompt
+    user_model_config = ChatGPTConfig(temperature=0.0)
+
+    function_list = function_list
+    assistant_model_config = FunctionCallingConfig.from_openai_function_list(
+        function_list=function_list,
+        kwargs=dict(temperature=0.0),
+    )
+
+    role_play_session = RolePlaying(
+        assistant_role_name=assistant_role_name,
+        user_role_name=user_role_name,
+        assistant_agent_kwargs=dict(
+            model_type=model_type,
+            model_config=assistant_model_config,
+            function_list=function_list,
+        ),
+        user_agent_kwargs=dict(
+            model_type=model_type,
+            model_config=user_model_config,
+        ),
+        task_prompt=task_prompt,
+        with_task_specify=False,
+    )
+
+    print(
+        Fore.GREEN +
+        f"AI Assistant sys message:\n{role_play_session.assistant_sys_msg}\n")
+    print(Fore.BLUE +
+          f"AI User sys message:\n{role_play_session.user_sys_msg}\n")
+
+    print(Fore.YELLOW + f"Original task prompt:\n{task_prompt}\n")
+    print(
+        Fore.CYAN +
+        f"Specified task prompt:\n{role_play_session.specified_task_prompt}\n")
+    print(Fore.RED + f"Final task prompt:\n{role_play_session.task_prompt}\n")
+
+    n = 0
+    input_assistant_msg, _ = role_play_session.init_chat()
+    while n < chat_turn_limit:
+        n += 1
+        assistant_response, user_response = role_play_session.step(
+            input_assistant_msg)
+
+        if assistant_response.terminated:
+            print(Fore.GREEN +
+                  ("AI Assistant terminated. Reason: "
+                   f"{assistant_response.info['termination_reasons']}."))
+            break
+        if user_response.terminated:
+            print(Fore.GREEN +
+                  ("AI User terminated. "
+                   f"Reason: {user_response.info['termination_reasons']}."))
+            break
+
+        # Print output from the user
+        print_text_animated(Fore.BLUE +
+                            f"AI User:\n\n{user_response.msg.content}\n")
+
+        # Print output from the assistant, including any function
+        # execution information
+        print_text_animated(Fore.GREEN + "AI Assistant:")
+        called_functions: List[
+            FunctionCallingRecord] = assistant_response.info[
+                'called_functions']
+        for func_record in called_functions:
+            print_text_animated(f"{func_record}")
+        print_text_animated(f"{assistant_response.msg.content}\n")
+
+        if "CAMEL_TASK_DONE" in user_response.msg.content:
+            break
+
+        input_assistant_msg = assistant_response.msg
